app/requirements.txt
# General
requests

# Django
Django==4.0
django-timezone-field 
django-cors-headers

# DRF
django-filter
djangorestframework==3.13.1

# Databases
psycopg2-binary==2.9.1

# File control
Pillow
django-cleanup

# Prepocessing
flashtext
cyrtranslit

# Server
gunicorn==20.1.0
